refactor: move Render into editor_widget_render.go
Relocate Render and its private helpers (diagStyleAt, renderLineToScreen,
screenCell) into editor_widget_render.go. DiagnosticAt stays in the core
file since it is an app-layer diagnostic lookup, not a render helper. Main
file drops to ~620 lines. Pure mechanical move, no behavior change.
